% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cte2list.R
\name{cte2list}
\alias{cte2list}
\title{Assign tables from a CTE SQL query to a list}
\usage{
cte2list()
}
\value{
No return value, called for side effects
}
\description{
Reads the currently open query from the RStudio API and returns
  a list of data frames for each CTE query.
}
