% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cte2views.R
\name{cte2views}
\alias{cte2views}
\title{Creates a View() output for each expression in a CTE SQL query}
\usage{
cte2views()
}
\value{
No return value, called for side effects
}
\description{
Reads the currently open query from the RStudio API and creates
  a View() tab for each expression in a CTE SQL query.
}
